Rename proxy.py (a support library) to less generic pac4cli.py
